reporting-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.7
Pillow
python-dotenv==1.0.0
PyYAML==6.0.1
